Updated install and other functions for automation

+++ installer-util.sh
@@ -1,17 +1,17 @@
 #!/usr/bin/env bash
 
-### Menu option variables
+# Menu option variables
 INPUT=/tmp/menu.sh.$$
 menuselection=""
 
-### Arrays
+# Arrays
 declare -a mainmenuoptions=('Installation' 'Update' 'Utilities' 'Exit')
 declare -a installmenuoptions=('Standard Install' 'Dev Test Prereqs' 'Dev Test Install' 'Return' 'Exit')
 declare -a updatemenuoptions=('Standard Update' 'Backup and Update' 'Force Update' 'Return' 'Exit')
 declare -a utilitymenuoptions=('Backup' 'Restore' 'Renew Certs' 'Import Certs' 'Edit UWSGI config' 'Add Fail2ban - Use at your own risk' 'Run Server Troubleshooter' 'Return' 'Exit')
 declare -a cfgfiles=('InputAndError.cfg' 'MiscFunctions.cfg' 'SystemInfoFunctions.cfg' 'UserInput.cfg' 'NetworkFunctions.cfg' 'InstallFunctions.cfg' 'DatabaseFunctions.cfg' 'CertificateFunctions.cfg' 'ConfigAndServiceFunctions.cfg' 'UpdateRestoreFunctions.cfg' 'TroubleshootingFunctions.cfg' 'ParentFunctions.cfg')
 
-### Script Info variables
+# Script Info variables
 REPO_OWNER="ninjamonkey198206"
 BRANCH="develop-installer-update"
 SCRIPT_VERSION="69"
@@ -22,7 +22,7 @@ SCRIPTS_REPO_URL="https://github.com/amidaware/community-scripts.git"
 FRONTEND_URL="https://github.com/amidaware/tacticalrmm-web/releases/download/v${WEB_VERSION}/${webtar}"
 THIS_SCRIPT=$(readlink -f "$0")
 
-### Misc info variables
+# Misc info variables
 INSTALL_TYPE="install"
 UPDATE_TYPE="standard"
 SCRIPTS_DIR='/opt/trmm-community-scripts'
@@ -30,7 +30,7 @@ PYTHON_VER='3.10.4'
 SETTINGS_FILE='/rmm/api/tacticalrmm/tacticalrmm/settings.py'
 LATEST_SETTINGS_URL="https://raw.githubusercontent.com/${REPO_OWNER}/tacticalrmm/${BRANCH}/api/tacticalrmm/tacticalrmm/settings.py"
 
-### Variables required for automated install
+# Variables required for automated install
 autoinstall=""
 rmmhost=""
 frontendhost=""
@@ -40,9 +40,9 @@ rootdomain=""
 sslcacert=""
 sslkey=""
 sslcert=""
+trmmuser=""
 
-
-### Get cfg files function
+# Get cfg files function
 getCfgFiles()
 {
 	if [ ! -f "$PWD/script-cfg/$2" ]; then
@@ -50,18 +50,18 @@ getCfgFiles()
 	fi
 }
 
-### Check if directory exists, if not, create
+# Check if directory exists, if not, create
 if [ ! -d $PWD/script-cfg ]; then
 	mkdir $PWD/script-cfg
 fi
 
-### Get cfg files
+# Get cfg files
 for i in "${cfgfiles[@]}"
 do
 	getCfgFiles "$CFG_URL" "$i";
 done
 
-### Import functions
+# Import functions
 . $PWD/script-cfg/InputAndError.cfg
 . $PWD/script-cfg/MiscFunctions.cfg
 . $PWD/script-cfg/SystemInfoFunctions.cfg
@@ -75,10 +75,10 @@ done
 . $PWD/script-cfg/TroubleshootingFunctions.cfg
 . $PWD/script-cfg/ParentFunctions.cfg
 
-### Get commandline input function
+# Get commandline input function
 getCommandLineArgs()
 {
-	while getopts "auto:api:branch:ca:cert:domain:email:help:key:mesh:repo:rmm:" option; do
+	while getopts "auto:api:branch:ca:cert:domain:email:help:key:mesh:repo:rmm:username:" option; do
 		case $option in
       		auto ) autoinstall="1"
 				INSTALL_TYPE="${OPTARG}";;
@@ -94,6 +94,7 @@ getCommandLineArgs()
 			mesh ) meshhost="${OPTARG}";;
 			repo ) REPO_OWNER="${OPTARG}";;
 			rmm ) frontendhost="${OPTARG}";;
+			username ) trmmuser="${OPTARG}";;
 	     	\?) echo -e "Error: Invalid option"
 				clear -x
 				exit 1;;
@@ -101,19 +102,19 @@ getCommandLineArgs()
 	done
 
 	if [ "$autoinstall" == "1" ]; then
-		if [ -z "$rmmhost" ] || [ -z "$sslcacert" ] || [ -z "$sslcert" ] || [ -z "$rootdomain" ] || [ -z "$letsemail" ] || [ -z "$sslkey" ] || [ -z "$meshhost" ] || [ -z "$frontendhost" ]; then
+		if [ -z "$rmmhost" ] || [ -z "$sslcacert" ] || [ -z "$sslcert" ] || [ -z "$rootdomain" ] || [ -z "$sslkey" ] || [ -z "$meshhost" ] || [ -z "$frontendhost" ] || [ -z "$trmmuser" ] || [ -z "$letsemail" ]; then
 			echo -e "Error: To perform an automated installation, you must provide all required information."
 			echo -e "\n"
-			echo -e "api host, mesh host, rmm host, root domain, email address, CA cert path, Cert path, and Private key path are all required."
+			echo -e "api host, mesh host, rmm host, root domain, email address, CA cert path, Cert path, Private key path, and T-RMM username are all required."
 			echo -e "\n"
-			echo -e "Run ./$THIS_SCRIPT -h for further details."
+			echo -e "Run .$THIS_SCRIPT -h for further details."
 			clear -x
 			exit 1
 		elif [ "$INSTALL_TYPE" == "devprep" ] || [ "$INSTALL_TYPE" == "devinstall" ]; then
 			if [ "$REPO_OWNER" == "amidaware" ] || [ "$BRANCH" == "master" ]; then
 				echo -e "Error: You've selected a developer installation type, but not changed the repo, branch, or both."
 				echo -e "\n"
-				echo -e "Run ./$THIS_SCRIPT -h for details on how to select them."
+				echo -e "Run .$THIS_SCRIPT -h for details on how to select them."
 				clear -x
 				exit 1
 			fi
@@ -123,16 +124,16 @@ getCommandLineArgs()
 	fi
 }				
 
-### Get commandline input
+# Get commandline input
 getCommandLineArgs;
 
-### Set colors
+# Set colors
 setColors;		# MiscFunctions
 
-### Gather OS info
+# Gather OS info
 getOSInfo;		# SystemInfoFunctions
 
-### Install script pre-reqs
+# Install script pre-reqs
 installPreReqs;		# InstallFunctions
 
 ### Check for new functions versions, include url, filename, and script name as variables
@@ -276,26 +277,29 @@ updateMenu()
 # Main menu
 mainMenu()
 {
-	until [ "$menuselection" = "0" ]; do
-		dialog --cr-wrap --clear --no-ok --no-cancel --backtitle "Tactical RMM Installation and Maintenance Utility" --title "Main Menu" --menu "Use the 'Up' and 'Down' keys to navigate, and the 'Enter' key to make your selections." 0 0 0 \
-			1 "${mainmenuoptions[0]}" \
-      		2 "${mainmenuoptions[1]}" \
-      		3 "${mainmenuoptions[2]}" \
-			4 "${mainmenuoptions[3]}" 2>"${INPUT}"
+	if [ "$autoinstall" == "1" ]; then
+		mainInstall;
+	else
+		until [ "$menuselection" = "0" ]; do
+			dialog --cr-wrap --clear --no-ok --no-cancel --backtitle "Tactical RMM Installation and Maintenance Utility" --title "Main Menu" --menu "Use the 'Up' and 'Down' keys to navigate, and the 'Enter' key to make your selections." 0 0 0 \
+				1 "${mainmenuoptions[0]}" \
+      			2 "${mainmenuoptions[1]}" \
+      			3 "${mainmenuoptions[2]}" \
+				4 "${mainmenuoptions[3]}" 2>"${INPUT}"
 
-		menuselection=$(<"${INPUT}")
+			menuselection=$(<"${INPUT}")
 
-		case $menuselection in
-			1 ) installMenu;;
-      		2 ) updateMenu;;
-      		3 )	utilityMenu;;
-			4 ) [ -f $INPUT ] && rm $INPUT
-				clear -x
-				exit;;
-			* ) derpDerp;;
-		esac
-	done
-
+			case $menuselection in
+				1 ) installMenu;;
+      			2 ) updateMenu;;
+      			3 )	utilityMenu;;
+				4 ) [ -f $INPUT ] && rm $INPUT
+					clear -x
+					exit;;
+				* ) derpDerp;;
+			esac
+		done
+	fi
 	return
 }
